Contract framework for validation
With aggressive use of 3.6+ features, an aspect programming based
framework for validation can be created.  The basic idea is that python
type annotations offer a nice representation for automatically
validating types.  The types here are validus types.

There is a minor issue between validus using 'isint' and 'isfloat'
which leads to type conflict with intrinsic types.  These validus names
should probably changed to isinteger and isreal, but this is not done to
avoid compatibility issues.

This is basically the framework described by David Beazley in his PyCon
Israel 2017 talk only using Validus for the checking (minor fixes and
improvements were made here and there to make it more robust.